requirements.txt
flask==2.2.3
pytesseract==0.3.10
pillow==9.5.0
python-dotenv==1.0.0
requests==2.28.2

